fix(security): make all pages server components
Only dynamic server components have the csp nonce added to them, client rendered pages must be wrapped.
